+++ CHANGELOG.md
@@ -12,7 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated 
 ### Removed 
 ### Fixed 
-### Security  -->
+### Security   -->
+
+## [2.1.04] - 2025-04-01
+### Added 
+- `Griptape Util: Switch Node` - This node lets you choose an output from multiple inputs.
+
+  ![switch node](docs/images/switch_node.png)
 
 ## [2.1.03] - 2025-03-01
 ### Fixed